Compute cell-modulated DCT in fragment shader
With set uniforms specifying the pixel and cell size in texture
coordinates (0,1), find the DCT for this pixel of the cell, which
corresponds to the spatial frequency of the cell at that cell UV.

The shader cannot compile right now, because it cannot loop to a
runtime-known maximum. This can be solved by recompiling the shaders
each time the image or cell size are reselected, setting new constants
for the pixel and cell sizes, or switching to WebGPU for compute
shaders.