Immediately refresh UI on page load
Previously, the page would not launch the WebGL context and process the
selected image on page start until the user initiated another change
event by reselecting the input image (which might already be cached as
selected between a refresh). This is fixed simply by calling the event
function once in the top level script.