Compute DCTs of glyphs in batches
Locally within the function, create a 2D Canvas context in order to draw
characters into a grid, buffer that canvas as a texture into the WebGL
context, and pass it through the DCT shader to compute the DCT of each
glyph for the global cell size.

Created a helper function refreshAndUseDctShader() to recompile the DCT
shader only when the global cell size changes, and reconfigured the
other two shader constants pixelSize and cellSize to be passed as
uniforms.