Normalize DCTs
The DCT shader was computing an unnormalized transform: each component
of a block was in the range [1, -1], for `cell_size^2` components, but
as the value is written to the R component of a pixel, the shader
normalizes the result to [0, 1], clipping almost every pixel. The DCT
values are now normalized according to the cell size, meaning they can
only be compared for DCTs computed over the same cell size (which was
already expected).